add spec and tests
This commit adds rspec tests as well as the tests directory for testing
the module directly.

The commit also includes Gemfile, mostly gems for testing the module, a
Rakefile to easily run tests and a Modulefile to be able to upload the
module to Puppet forge.

RSpec tests are currently very naive and should be extended and should
include more fine grained parameters.